release/blue

+++ scripts/linux/after-uninstall.sh
@@ -1,15 +1,36 @@
 #!/bin/bash
 
-echo 'Removing configs'
-rm -rf ~/.config/Sentinel\ dVPN
+# Get the root user
+if [ $SUDO_USER ];
+	then getSudoUser=$SUDO_USER;
+	else getSudoUser=`whoami`;
+fi
 
-echo 'Removing sentinelcli'
-rm -rf /usr/local/bin/sentinelcli
+appDirectory=/home/$getSudoUser/.config/Sentinel\ dVPN/;
+if [ -d "$appDirectory" ]; then
+    echo 'Removing the configs'
+    rm -rf "$appDirectory";
+fi
 
-echo 'Removing the daemon'
-systemctl stop sentinelcli-vpn.service
-systemctl disable sentinelcli-vpn.service
-rm /etc/systemd/system/sentinelcli-vpn.service
-# rm /etc/systemd/system/multi-user.target.wants/sentinelcli-vpn.service
-systemctl daemon-reload
-systemctl reset-failed
+sentinelcliDirectory=/home/$getSudoUser/.sentinelcli-dvpn-desktop/;
+if [ -d "$sentinelcliDirectory" ]; then
+    echo 'Removing the sentinelcli configs'
+    rm -rf "$sentinelcliDirectory";
+fi
+
+vpnDaemonFile=/etc/systemd/system/sentinelcli-vpn.service
+if [ -f "$vpnDaemonFile" ]; then
+  echo 'Removing the daemon'
+  systemctl stop sentinelcli-vpn.service
+  systemctl disable sentinelcli-vpn.service
+  rm "$vpnDaemonFile"
+  # rm /etc/systemd/system/multi-user.target.wants/sentinelcli-vpn.service
+  systemctl daemon-reload
+  systemctl reset-failed
+fi
+
+keyringPid=$(pgrep -f "sentinelcli.+\--with-keyring") || keyringPid=0
+if [ $keyringPid -gt 0 ]; then
+  echo 'Killing a sentinelcli keyring process'
+  kill $keyringPid
+fi
